Move translation to deprecated module
