+++ src/configuration/legacy_translation.rs
@@ -1,0 +1,547 @@
+use super::{
+    Action, ConditionalData, DataItem, DataType, DenyOperation, FailOperation, GrpcOperation,
+    HeadersOperation, HeadersTarget, Operation, TypedAction,
+};
+
+pub(super) mod ratelimit {
+    use super::*;
+
+    fn escape_cel_string(s: &str) -> String {
+        s.replace('\\', "\\\\")
+            .replace('"', "\\\"")
+            .replace('\n', "\\n")
+            .replace('\r', "\\r")
+            .replace('\t', "\\t")
+    }
+
+    const RATELIMIT_KNOWN_ATTRS: [&str; 2] = ["ratelimit.domain", "ratelimit.hits_addend"];
+
+    fn is_ratelimit_known_attr(item: &DataItem) -> bool {
+        let key = match &item.item {
+            DataType::Static(s) => s.key.as_str(),
+            DataType::Expression(e) => e.key.as_str(),
+        };
+        RATELIMIT_KNOWN_ATTRS.contains(&key)
+    }
+
+    fn find_ratelimit_known_attr_cel(
+        conditional_data: &[ConditionalData],
+        attr_key: &str,
+    ) -> Option<String> {
+        for cd in conditional_data {
+            for item in &cd.data {
+                match &item.item {
+                    DataType::Static(s) if s.key == attr_key => {
+                        return Some(format!(r#""{}""#, escape_cel_string(&s.value)));
+                    }
+                    DataType::Expression(e) if e.key == attr_key => {
+                        return Some(e.value.clone());
+                    }
+                    _ => {}
+                }
+            }
+        }
+        None
+    }
+
+    fn build_ratelimit_descriptor_entry_cel(item: &DataItem) -> String {
+        let (key, value_cel) = match &item.item {
+            DataType::Static(s) => (
+                s.key.as_str(),
+                format!(r#""{}""#, escape_cel_string(&s.value)),
+            ),
+            DataType::Expression(e) => (e.key.as_str(), format!("string({})", e.value)),
+        };
+
+        format!(
+            r#"envoy.extensions.common.ratelimit.v3.RateLimitDescriptor.Entry {{ key: "{}", value: {} }}"#,
+            escape_cel_string(key),
+            value_cel,
+        )
+    }
+
+    fn build_ratelimit_entry_list_cel(cd: &ConditionalData) -> Option<String> {
+        let entries: Vec<String> = cd
+            .data
+            .iter()
+            .filter(|item| !is_ratelimit_known_attr(item))
+            .map(build_ratelimit_descriptor_entry_cel)
+            .collect();
+
+        if entries.is_empty() {
+            return None;
+        }
+
+        let entries_list = format!("[{}]", entries.join(", "));
+
+        if cd.predicates.is_empty() {
+            Some(entries_list)
+        } else {
+            let wrapped: Vec<String> = cd.predicates.iter().map(|p| format!("({})", p)).collect();
+            let predicate_cel = wrapped.join(" && ");
+            Some(format!("(({}) ? {} : [])", predicate_cel, entries_list))
+        }
+    }
+
+    fn build_ratelimit_descriptors_cel(conditional_data: &[ConditionalData]) -> Option<String> {
+        let entry_parts: Vec<String> = conditional_data
+            .iter()
+            .filter_map(build_ratelimit_entry_list_cel)
+            .collect();
+
+        if entry_parts.is_empty() {
+            return None;
+        }
+
+        let combined_entries = entry_parts.join(" + ");
+        Some(format!(
+            "envoy.extensions.common.ratelimit.v3.RateLimitDescriptor {{ entries: {} }}",
+            combined_entries
+        ))
+    }
+
+    fn build_ratelimit_message_builder(
+        scope: &str,
+        conditional_data: &[ConditionalData],
+        request_data: &[((String, String), String)],
+    ) -> String {
+        let domain_cel = find_ratelimit_known_attr_cel(conditional_data, "ratelimit.domain")
+            .unwrap_or_else(|| format!(r#""{}""#, escape_cel_string(scope)));
+
+        let hits_addend_cel =
+            find_ratelimit_known_attr_cel(conditional_data, "ratelimit.hits_addend")
+                .unwrap_or_else(|| "1u".to_string());
+
+        let mut descriptors = vec![];
+
+        if let Some(desc) = build_ratelimit_descriptors_cel(conditional_data) {
+            descriptors.push(desc);
+        }
+
+        if !request_data.is_empty() {
+            let request_data_entries: Vec<String> = request_data
+                .iter()
+                .map(|((domain, field), value_expr)| {
+                    let key = if domain.is_empty() || domain == "metrics.labels" {
+                        field.clone()
+                    } else {
+                        format!("{}.{}", domain, field)
+                    };
+                    format!(
+                        r#"envoy.extensions.common.ratelimit.v3.RateLimitDescriptor.Entry {{ key: "{}", value: {} }}"#,
+                        escape_cel_string(&key),
+                        value_expr
+                    )
+                })
+                .collect();
+
+            if !request_data_entries.is_empty() {
+                descriptors.push(format!(
+                    "envoy.extensions.common.ratelimit.v3.RateLimitDescriptor {{ entries: [{}] }}",
+                    request_data_entries.join(", ")
+                ));
+            }
+        }
+
+        let descriptors_cel = if descriptors.is_empty() {
+            "[]".to_string()
+        } else {
+            format!("[{}]", descriptors.join(", "))
+        };
+
+        format!(
+            r#"envoy.service.ratelimit.v3.RateLimitRequest {{
+    domain: {},
+    hits_addend: {},
+    descriptors: {}
+}}"#,
+            domain_cel, hits_addend_cel, descriptors_cel,
+        )
+    }
+
+    fn build_descriptor_predicate(conditional_data: &[ConditionalData]) -> String {
+        if conditional_data.is_empty() {
+            return "true".to_string();
+        }
+        if conditional_data.iter().any(|cd| cd.predicates.is_empty()) {
+            return "true".to_string();
+        }
+
+        let block_predicates: Vec<String> = conditional_data
+            .iter()
+            .map(|cd| {
+                if cd.predicates.len() == 1 {
+                    cd.predicates[0].clone()
+                } else {
+                    let wrapped: Vec<String> =
+                        cd.predicates.iter().map(|p| format!("({})", p)).collect();
+                    format!("({})", wrapped.join(" && "))
+                }
+            })
+            .collect();
+
+        if block_predicates.len() == 1 {
+            block_predicates[0].clone()
+        } else {
+            block_predicates.join(" || ")
+        }
+    }
+
+    fn build_action_predicate(action_predicates: &[String]) -> String {
+        if action_predicates.is_empty() {
+            "true".to_string()
+        } else if action_predicates.len() == 1 {
+            action_predicates[0].clone()
+        } else {
+            let wrapped: Vec<String> = action_predicates
+                .iter()
+                .map(|p| format!("({})", p))
+                .collect();
+            wrapped.join(" && ")
+        }
+    }
+
+    fn build_ratelimit_predicate(
+        action_predicates: &[String],
+        conditional_data: &[ConditionalData],
+    ) -> String {
+        let action_pred = build_action_predicate(action_predicates);
+        let conditional_pred = build_descriptor_predicate(conditional_data);
+
+        if action_pred == "true" && conditional_pred == "true" {
+            "true".to_string()
+        } else if action_pred == "true" {
+            conditional_pred
+        } else if conditional_pred == "true" {
+            action_pred
+        } else {
+            format!("({}) && ({})", action_pred, conditional_pred)
+        }
+    }
+
+    fn build_ratelimit_on_reply(name: &str) -> Vec<TypedAction> {
+        vec![
+            TypedAction {
+                predicate: format!("{}.overall_code == 2", name),
+                terminal: true,
+                operation: Operation::Deny(DenyOperation {
+                    deny_with: format!(
+                        r#"DenyResponse{{status: 429u, headers: {}.response_headers_to_add, body: "Too Many Requests\n"}}"#,
+                        name
+                    ),
+                }),
+            },
+            TypedAction {
+                predicate: format!("{}.overall_code == 1", name),
+                terminal: false,
+                operation: Operation::Headers(HeadersOperation {
+                    target: HeadersTarget::Response,
+                    headers: format!("{}.response_headers_to_add", name),
+                }),
+            },
+            TypedAction {
+                predicate: format!("{}.overall_code != 1 && {}.overall_code != 2", name, name),
+                terminal: true,
+                operation: Operation::Fail(FailOperation {
+                    log_message: format!("Unknown rate limit response code from {}", name),
+                }),
+            },
+        ]
+    }
+
+    #[deprecated(note = "temporary translation for legacy ratelimit configuration")]
+    pub(crate) fn translate_legacy_ratelimit_to_typed(
+        action: &Action,
+        request_data: &[((String, String), String)],
+    ) -> TypedAction {
+        const RESPONSE_VAR: &str = "ratelimit_response";
+
+        let message_builder =
+            build_ratelimit_message_builder(&action.scope, &action.conditional_data, request_data);
+
+        let predicate = build_ratelimit_predicate(&action.predicates, &action.conditional_data);
+
+        let on_reply = build_ratelimit_on_reply(RESPONSE_VAR);
+
+        TypedAction {
+            predicate,
+            terminal: false,
+            operation: Operation::Grpc(GrpcOperation {
+                var: RESPONSE_VAR.to_string(),
+                service: action.service.clone(),
+                message_builder,
+                on_reply,
+            }),
+        }
+    }
+
+    #[cfg(test)]
+    #[allow(deprecated)]
+    mod tests {
+        use crate::configuration::{ExpressionItem, StaticItem};
+
+        use super::*;
+
+        #[test]
+        fn test_build_descriptor_predicate_empty() {
+            let conditional_data = vec![];
+            assert_eq!(build_descriptor_predicate(&conditional_data), "true");
+        }
+
+        #[test]
+        fn test_build_descriptor_predicate_unconditional() {
+            let conditional_data = vec![ConditionalData {
+                predicates: vec![],
+                data: vec![DataItem {
+                    item: DataType::Static(StaticItem {
+                        key: "limit".to_string(),
+                        value: "10".to_string(),
+                    }),
+                }],
+            }];
+            assert_eq!(build_descriptor_predicate(&conditional_data), "true");
+        }
+
+        #[test]
+        fn test_build_descriptor_predicate_single_predicate() {
+            let conditional_data = vec![ConditionalData {
+                predicates: vec!["auth.identity.user == 'alice'".to_string()],
+                data: vec![],
+            }];
+            assert_eq!(
+                build_descriptor_predicate(&conditional_data),
+                "auth.identity.user == 'alice'"
+            );
+        }
+
+        #[test]
+        fn test_build_descriptor_predicate_multiple_predicates_single_block() {
+            let conditional_data = vec![ConditionalData {
+                predicates: vec![
+                    "auth.identity.user == 'alice'".to_string(),
+                    "request.method == 'POST'".to_string(),
+                ],
+                data: vec![],
+            }];
+            assert_eq!(
+                build_descriptor_predicate(&conditional_data),
+                "((auth.identity.user == 'alice') && (request.method == 'POST'))"
+            );
+        }
+
+        #[test]
+        fn test_build_descriptor_predicate_multiple_blocks() {
+            let conditional_data = vec![
+                ConditionalData {
+                    predicates: vec!["auth.identity.role == 'admin'".to_string()],
+                    data: vec![],
+                },
+                ConditionalData {
+                    predicates: vec!["auth.identity.role == 'user'".to_string()],
+                    data: vec![],
+                },
+            ];
+            assert_eq!(
+                build_descriptor_predicate(&conditional_data),
+                "auth.identity.role == 'admin' || auth.identity.role == 'user'"
+            );
+        }
+
+        #[test]
+        fn test_build_descriptor_predicate_mixed_conditional_unconditional() {
+            let conditional_data = vec![
+                ConditionalData {
+                    predicates: vec!["auth.identity.role == 'admin'".to_string()],
+                    data: vec![],
+                },
+                ConditionalData {
+                    predicates: vec![],
+                    data: vec![],
+                },
+            ];
+            assert_eq!(build_descriptor_predicate(&conditional_data), "true");
+        }
+
+        #[test]
+        fn test_build_ratelimit_on_reply_structure() {
+            let on_reply = build_ratelimit_on_reply("rl_response");
+
+            assert_eq!(on_reply.len(), 3);
+
+            assert_eq!(on_reply[0].predicate, "rl_response.overall_code == 2");
+            assert!(on_reply[0].terminal);
+            assert!(matches!(on_reply[0].operation, Operation::Deny(_)));
+
+            assert_eq!(on_reply[1].predicate, "rl_response.overall_code == 1");
+            assert!(!on_reply[1].terminal);
+            assert!(matches!(on_reply[1].operation, Operation::Headers(_)));
+
+            assert_eq!(
+                on_reply[2].predicate,
+                "rl_response.overall_code != 1 && rl_response.overall_code != 2"
+            );
+            assert!(on_reply[2].terminal);
+            assert!(matches!(on_reply[2].operation, Operation::Fail(_)));
+        }
+
+        #[test]
+        fn test_build_ratelimit_on_reply_deny_operation() {
+            let on_reply = build_ratelimit_on_reply("test_var");
+
+            assert!(matches!(&on_reply[0].operation,
+                Operation::Deny(deny_op) if
+                    deny_op.deny_with.contains("DenyResponse") &&
+                    deny_op.deny_with.contains("status: 429u") &&
+                    deny_op.deny_with.contains("test_var.response_headers_to_add") &&
+                    deny_op.deny_with.contains("Too Many Requests")
+            ));
+        }
+
+        #[test]
+        fn test_build_ratelimit_on_reply_headers_operation() {
+            let on_reply = build_ratelimit_on_reply("my_rl");
+
+            assert!(matches!(&on_reply[1].operation,
+                Operation::Headers(headers_op) if
+                    matches!(headers_op.target, HeadersTarget::Response) &&
+                    headers_op.headers == "my_rl.response_headers_to_add"
+            ));
+        }
+
+        #[test]
+        fn test_build_ratelimit_on_reply_fail_operation() {
+            let on_reply = build_ratelimit_on_reply("rate_limit");
+
+            assert!(matches!(&on_reply[2].operation,
+                Operation::Fail(fail_op) if
+                    fail_op.log_message.contains("Unknown rate limit response code") &&
+                    fail_op.log_message.contains("rate_limit")
+            ));
+        }
+
+        #[test]
+        fn test_translate_legacy_ratelimit_simple() {
+            let action = Action {
+                service: "limitador".to_string(),
+                scope: "my-ratelimit".to_string(),
+                predicates: vec![],
+                conditional_data: vec![],
+                sources: vec![],
+            };
+            let request_data = vec![];
+
+            let typed = translate_legacy_ratelimit_to_typed(&action, &request_data);
+
+            assert_eq!(typed.predicate, "true");
+            assert!(!typed.terminal);
+
+            assert!(matches!(&typed.operation,
+                Operation::Grpc(grpc_op) if
+                    grpc_op.var == "ratelimit_response" &&
+                    grpc_op.service == "limitador" &&
+                    grpc_op.message_builder.contains(r#"domain: "my-ratelimit""#) &&
+                    grpc_op.on_reply.len() == 3
+            ));
+        }
+
+        #[test]
+        fn test_translate_legacy_ratelimit_with_conditional_data() {
+            let action = Action {
+                service: "limitador".to_string(),
+                scope: "my-ratelimit".to_string(),
+                predicates: vec![],
+                conditional_data: vec![ConditionalData {
+                    predicates: vec!["auth.identity.user == 'alice'".to_string()],
+                    data: vec![DataItem {
+                        item: DataType::Static(StaticItem {
+                            key: "tier".to_string(),
+                            value: "gold".to_string(),
+                        }),
+                    }],
+                }],
+                sources: vec![],
+            };
+            let request_data = vec![];
+
+            let typed = translate_legacy_ratelimit_to_typed(&action, &request_data);
+
+            assert_eq!(typed.predicate, "auth.identity.user == 'alice'");
+            assert!(matches!(&typed.operation,
+                Operation::Grpc(grpc_op) if
+                    grpc_op.message_builder.contains("auth.identity.user == 'alice'") &&
+                    grpc_op.message_builder.contains(r#"key: "tier""#) &&
+                    grpc_op.message_builder.contains(r#"value: "gold""#)
+            ));
+        }
+
+        #[test]
+        fn test_translate_legacy_ratelimit_with_request_data() {
+            let action = Action {
+                service: "limitador".to_string(),
+                scope: "default".to_string(),
+                predicates: vec![],
+                conditional_data: vec![],
+                sources: vec![],
+            };
+            let request_data = vec![(
+                ("".to_string(), "env".to_string()),
+                r#""production""#.to_string(),
+            )];
+
+            let typed = translate_legacy_ratelimit_to_typed(&action, &request_data);
+
+            assert!(matches!(&typed.operation,
+                Operation::Grpc(grpc_op) if
+                    grpc_op.message_builder.contains(r#"key: "env""#) &&
+                    grpc_op.message_builder.contains(r#""production""#)
+            ));
+        }
+
+        #[test]
+        fn test_translate_legacy_ratelimit_full() {
+            let action = Action {
+                service: "limitador".to_string(),
+                scope: "rlp-full".to_string(),
+                predicates: vec![],
+                conditional_data: vec![
+                    ConditionalData {
+                        predicates: vec!["auth.identity.role == 'admin'".to_string()],
+                        data: vec![DataItem {
+                            item: DataType::Static(StaticItem {
+                                key: "tier".to_string(),
+                                value: "gold".to_string(),
+                            }),
+                        }],
+                    },
+                    ConditionalData {
+                        predicates: vec![],
+                        data: vec![DataItem {
+                            item: DataType::Expression(ExpressionItem {
+                                key: "method".to_string(),
+                                value: "request.method".to_string(),
+                            }),
+                        }],
+                    },
+                ],
+                sources: vec![],
+            };
+            let request_data = vec![(
+                ("".to_string(), "env".to_string()),
+                r#""production""#.to_string(),
+            )];
+
+            let typed = translate_legacy_ratelimit_to_typed(&action, &request_data);
+
+            assert_eq!(typed.predicate, "true");
+
+            assert!(matches!(&typed.operation,
+                Operation::Grpc(grpc_op) if
+                    grpc_op.message_builder.contains(r#"domain: "rlp-full""#) &&
+                    grpc_op.message_builder.contains("auth.identity.role == 'admin'") &&
+                    grpc_op.message_builder.contains(r#"key: "tier""#) &&
+                    grpc_op.message_builder.contains(r#"key: "method""#) &&
+                    grpc_op.message_builder.contains(r#"key: "env""#)
+            ));
+        }
+    }
+}
